feat: Update Replace Unwanted Characters plugin to v1.1.1
- 1.1.1 (1bb8f5e)
- refactor(settings_ui): ReplaceUnwantedCharactersOptionsPage.save is too complex (18) (MC0001) (bb8fecf)
- refactor(settings_ui): ReplaceUnwantedCharactersOptionsPage.save is too complex (18) (MC0001) (b7727a3)
- refactor: remove redundant blank lines in __init__.py and options.py (38429ef)
- fix(replace_unwanted_characters): add tooltips for configuration options (98e0c61)
- Merge branch 'foobar' (d982268)
- fix: Simplify config setting retrieval by removing redundant try-except blocks (f051018)
- Update README.md (2aaaede)
- refactor: Move constants and options to separate files for better organization (e9eadbe)
