+++ plugins/replace_unwanted_characters/constants.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+
+
+
+DEFAULT_TAGS = ["album", "artist", "title", "albumartist", "releasetype", "label"]
+
+DEFAULT_CHAR_MAPPING = {
+    ":": "∶", "/": "⁄", "*": "∗", "?": "？", '"': '″',
+    '\\': '⧵', '.': '․', '|': 'ǀ', '<': '‹', '>': '›'
+}
